+++ src/features/Search/utils.test.ts
@@ -1,5 +1,5 @@
 import { searchableItems } from 'src/__data__/searchableItems';
-import { SearchableItem } from 'src/features/Search/refinedSearch';
+import { SearchableItem } from 'src/features/Search/search.interfaces';
 import { separateResultsByEntity } from './utils';
 
 const data = searchableItems as SearchableItem[];
