+++ src/features/Search/utils.ts
@@ -2,28 +2,7 @@ import DomainIcon from 'src/assets/addnewmenu/domain.svg';
 import LinodeIcon from 'src/assets/addnewmenu/linode.svg';
 import NodebalIcon from 'src/assets/addnewmenu/nodebalancer.svg';
 import VolumeIcon from 'src/assets/addnewmenu/volume.svg';
-import { SearchableItem } from 'src/features/Search/refinedSearch';
-
-export interface SearchResults {
-  combinedResults: SearchableItem[];
-  searchResultsByEntity: SearchResultsByEntity;
-}
-
-export interface SearchResultsByEntity {
-  linodes: SearchableItem[];
-  volumes: SearchableItem[];
-  nodebalancers: SearchableItem[];
-  domains: SearchableItem[];
-  images: SearchableItem[];
-}
-
-export const emptyResults: SearchResultsByEntity = {
-  linodes: [],
-  nodebalancers: [],
-  volumes: [],
-  domains: [],
-  images: []
-};
+import { SearchableItem, SearchResultsByEntity } from './search.interfaces';
 
 export const iconMap = {
   LinodeIcon,
@@ -31,6 +10,14 @@ export const iconMap = {
   VolumeIcon,
   DomainIcon,
   default: LinodeIcon
+};
+
+export const emptyResults: SearchResultsByEntity = {
+  linodes: [],
+  volumes: [],
+  domains: [],
+  images: [],
+  nodebalancers: []
 };
 
 export const separateResultsByEntity = (
